Add information about suggested mission profile
It's not entirely clear, since kOS can't tolerate sub-folders, which
programs are mission-worthy and which are interesting for technique
demonstration. The README has been re-done to give a sense of the
mission profile recommended given the maturity of the existing
programs.

The "scratch" programs might be moved to a non-installation
location. I am unsure.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>